Create Project 1: Bike Bright
This is a modification of the code and hard wiring that we did for Nuit Blanche. Two light sensors activate two LED strings, blinking the strings when light approaches; the light sensors and strings are mounted to a bike at night, adding visibility depending on the brightness of the area. Cruising down a quiet, dark side street, the lights are solid, but when the sensors pick up more light (like at an intersection or if headlights are approaching the bike), the LED strings blink, making the bike more visible.